### 2025-09-02 #8908 @editorialbot
[PRE REVIEW]: Reyna: A minimal overhead polytopal discontinuous Galerkin finite element library
**Submitting author:** @mattevs24 (<a href="http://orcid.org/0009-0000-5950-3751">Matthew Evans</a>)
**Repository:** https://github.com/mattevs24/reyna
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.3.0
