### 2025-09-02 42967d5
Added the DOIs
