+++ paper/references.bib
@@ -6,7 +6,8 @@
   number={2},
   pages={117--137},
   year={1999},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1023/A:1019155918070},
 }
 
 @article{Sutton:2017,
@@ -17,7 +18,8 @@
   number={4},
   pages={1141--1159},
   year={2017},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s11075-016-0235-3},
 }
 
 @article{Calloo:2025,
@@ -26,7 +28,8 @@
   journal={Numerical Algorithms},
   pages={1--24},
   year={2025},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s11075-025-02081-1},
 }
 
 @article{Cangiani:2014,
@@ -37,7 +40,8 @@
   number={10},
   pages={2009--2041},
   year={2014},
-  publisher={World Scientific}
+  publisher={World Scientific},
+  doi={10.1142/S0218202514500146},
 }
 
 @article{Cangiani:2016,
@@ -48,7 +52,8 @@
   number={3},
   pages={699--725},
   year={2016},
-  publisher={EDP Sciences}
+  publisher={EDP Sciences},
+  doi={10.1051/m2an/2015059},
 }
 
 @article{Talischi:2012,
@@ -59,7 +64,8 @@
   number={3},
   pages={309--328},
   year={2012},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s00158-011-0706-z},
 }
 
 @article{Dong:2020,
@@ -70,5 +76,6 @@
   number={4},
   pages={1309--1337},
   year={2020},
-  publisher={EDP Sciences}
+  publisher={EDP Sciences},
+  doi={10.1051/m2an/2019047},
 }
