### 2025-09-04 #8908 @danielskatz
Thanks @mattevs24 - Due to the relatively small amount of code in the core library and the commit history, the editors will now discuss if it meets the [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) criterion for review by JOSS. You should hear back in a week or two.
