### 2025-10-02 #8908 @mattevs24
Hi @drvinceknight, I'm not sure how many you recommend but my suggestions are Owen Melia (@meliao), Andrew Giuliani (@andrewgiuliani), Lambert Theisen (@lamBOOO) and Maximilian Bittens (@baxmittens). These choices are based on their experience in partial differential equations and finite element methods of various sorts. I hope this helps!
