### 2025-10-22 af67371
Small optimisations with the voronoi postprocessing -- increased performance with numba optimisation
