### 2025-11-04 mattevs24/reyna#1 @meliao
JOSS review: questions about functionality
Hi, congratulations on creating this nice package. As part of [the JOSS review](https://github.com/openjournals/joss-reviews/issues/9292), I've read the documentation and source code, and I have the following questions.

 * How does a user access the solution pointwise on the domain? It looks like this computation is present in `plot_DG` and `DGFEM.errors`; exposing this directly as a method may be very useful to users.
 * The README mentions that `reyna` can handle nonlinear equations, but the `DGFEM.add_data()` method only allows specification of coefficient fields in the linear elliptic PDE formulation. How can the code handle nonlinear equations?
