### 2025-11-04 mattevs24/reyna#1 @mattevs24
@meliao Currenlty users are not able to access the solution pointwise on the domain. As you say, this can be done through the functions, but the user is not able to input their own choice of points. Exposing this could be useful and can be done. I'm happy to work on this.

Support for non-linear equations is possible, but requires more user input. The code is able to assemble the necessary matrices but doesn't solve the nonlinear equations directly the user has to iterate towards the solution themselves. A sensible solution to avoid any confusion is to remove this line from the README until full support is offered.
