### 2025-11-04 74e7d80
Small updates for user information.
