### 2025-11-04 04218ff
Small updates to paper.md.
