+++ paper/references.bib
@@ -79,3 +79,45 @@
   publisher={EDP Sciences},
   doi={10.1051/m2an/2019047},
 }
+
+@article{Dedner:2017,
+  title={The DUNE-FEM-DG module.},
+  author={Dedner, Andreas and Girke, Stefan and Kl{\"o}fkorn, Robert and Malkmus, Tobias},
+  journal={Archive of Numerical Software},
+  volume={5},
+  number={1},
+  pages={21--61},
+  year={2017}
+}
+
+@article{Badia:2018,
+  title={FEMPAR: An object-oriented parallel finite element framework},
+  author={Badia, Santiago and Mart{\'\i}n, Alberto F and Principe, Javier},
+  journal={Archives of Computational Methods in Engineering},
+  volume={25},
+  number={2},
+  pages={195--271},
+  year={2018},
+  publisher={Springer}
+}
+
+@article{Anderson:2021,
+  title={MFEM: A modular finite element methods library},
+  author={Anderson, Robert and Andrej, Julian and Barker, Andrew and Bramwell, Jamie and Camier, Jean-Sylvain and Cerveny, Jakub and Dobrev, Veselin and Dudouit, Yohann and Fisher, Aaron and Kolev, Tzanio and others},
+  journal={Computers \& Mathematics with Applications},
+  volume={81},
+  pages={42--74},
+  year={2021},
+  publisher={Elsevier}
+}
+
+@inproceedings{Prud:2012,
+  title={Feel++: A computational framework for galerkin methods and advanced numerical methods},
+  author={Prud’Homme, Christophe and Chabannes, Vincent and Doyeux, Vincent and Ismail, Mourad and Samake, Abdoulaye and Pena, Gon{\c{c}}alo},
+  booktitle={ESAIM: Proceedings},
+  volume={38},
+  pages={429--455},
+  year={2012},
+  organization={EDP Sciences}
+}
+
