### 2025-11-04 8a51194
Small update - The format wasn't correct with the citations.
