### 2025-11-05 5923365
Added the discontinuous example and updated the paper.md file with some examples in the meantime.
