+++ paper/references.bib
@@ -90,6 +90,17 @@
   year={2017}
 }
 
+@article{Dedner:2017,
+  title={The DUNE-FEM-DG module.},
+  author={Dedner, Andreas and Girke, Stefan and Kl{\"o}fkorn, Robert and Malkmus, Tobias},
+  journal={Archive of Numerical Software},
+  volume={5},
+  number={1},
+  pages={21--61},
+  year={2017},
+  doi={10.11588/ans.2017.1.28602}
+}
+
 @article{Badia:2018,
   title={FEMPAR: An object-oriented parallel finite element framework},
   author={Badia, Santiago and Mart{\'\i}n, Alberto F and Principe, Javier},
@@ -98,7 +109,8 @@
   number={2},
   pages={195--271},
   year={2018},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s11831-017-9244-1}
 }
 
 @article{Anderson:2021,
@@ -108,7 +120,8 @@
   volume={81},
   pages={42--74},
   year={2021},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={https://doi.org/10.1016/j.camwa.2020.06.009},
 }
 
 @inproceedings{Prud:2012,
@@ -118,6 +131,40 @@
   volume={38},
   pages={429--455},
   year={2012},
-  organization={EDP Sciences}
+  organization={EDP Sciences},
+  doi={10.1051/proc/201238024}
 }
 
+@article{Bangerth:2007,
+  title={deal. II—a general-purpose object-oriented finite element library},
+  author={Bangerth, Wolfgang and Hartmann, Ralf and Kanschat, Guido},
+  journal={ACM Transactions on Mathematical Software (TOMS)},
+  volume={33},
+  number={4},
+  pages={24--es},
+  year={2007},
+  publisher={ACM New York, NY, USA},
+  doi={10.1145/1268776.1268779}
+}
+
+@article{alnaes:2015,
+  title={The FEniCS project version 1.5},
+  author={Aln{\ae}s, Martin and Blechta, Jan and Hake, Johan and Johansson, August and Kehlet, Benjamin and Logg, Anders and Richardson, Chris and Ring, Johannes and Rognes, Marie E and Wells, Garth N},
+  journal={Archive of numerical software},
+  volume={3},
+  number={100},
+  year={2015},
+  doi={10.11588/ans.2015.100.20553}
+}
+
+@article{Rathgeber:2016,
+  title={Firedrake: automating the finite element method by composing abstractions},
+  author={Rathgeber, Florian and Ham, David A and Mitchell, Lawrence and Lange, Michael and Luporini, Fabio and McRae, Andrew TT and Bercea, Gheorghe-Teodor and Markall, Graham R and Kelly, Paul HJ},
+  journal={ACM Transactions on Mathematical Software (TOMS)},
+  volume={43},
+  number={3},
+  pages={1--27},
+  year={2016},
+  publisher={ACM New York, NY, USA},
+  doi={10.1145/2998441}
+}
