+++ paper/references.bib
@@ -168,3 +168,14 @@
   publisher={ACM New York, NY, USA},
   doi={10.1145/2998441}
 }
+
+@article{Ching:2022,
+title = {Quail: A lightweight open-source discontinuous Galerkin code in Python for teaching and prototyping},
+journal = {SoftwareX},
+volume = {17},
+pages = {100982},
+year = {2022},
+issn = {2352-7110},
+doi = {https://doi.org/10.1016/j.softx.2022.100982},
+author = {Eric J. Ching and Brett Bornhoft and Ali Lasemi and Matthias Ihme},
+}
