+++ reyna/DGFEM/two_dimensional/main.py
@@ -28,7 +28,7 @@ from reyna.DGFEM.two_dimensional.plotter import plot_DG
 
 class DGFEM:
     """
-    This class comtains all the methods required for the DGFEM solver.
+    This class contains all the methods required for the DGFEM solver.
 
     Attributes:
         geometry (DGFEMGeometry): The geometry associated with the computation domain for the problem.
@@ -42,7 +42,7 @@ class DGFEM:
         dirichlet_bcs (typing.Optional[typing.Callable[[np.ndarray], np.ndarray]]): The dirichlet boundary conditions.
 
         boundary_information (typing.Optional[BoundaryInformation]): The information associated with the boundary of the
-        computational domain and PDE coefficients.
+            computational domain and PDE coefficients.
         sigma_D (float): The global penalisation coefficient.
 
         orders (typing.Optional[np.ndarray]): The orders of the tensor-Legendre polynomials.
@@ -56,9 +56,9 @@ class DGFEM:
         L (typing.Optional[np.ndarray]): The global forcng vector (populated when .dgfem() method is called).
 
         element_reference_quadrature (typing.Optional[typing.Tuple[np.ndarray, np.ndarray]]): The 2d reference
-        quadrature rule.
+            quadrature rule.
         edge_reference_quadrature (typing.Optional[typing.Tuple[np.ndarray, np.ndarray]]): The 1d reference quadrature
-        rule.
+            rule.
 
     Methods:
         add_data(...): Add the function coefficients for the non-negative characteristic second order PDE.
@@ -67,7 +67,7 @@ class DGFEM:
         plot_DG(...): Plot the DGFEM solution.
 
     See Also:
-        - The reference and example notebooks for examples on how this class is used fully.
+        - See the reference and example notebooks for examples on how this class is used fully.
     """
 
     def __init__(self, geometry: DGFEMGeometry, polynomial_degree: int = 1):
@@ -129,17 +129,17 @@ class DGFEM:
 
         Args:
             advection: The advection component of the PDE. Must be able to handle inputs with dimension (N,2) and output
-            arrays with dimension (N,2). I.e. the N advection coefficients at the given N 2D input locations.
+                arrays with dimension (N,2). I.e. the N advection coefficients at the given N 2D input locations.
             diffusion: The diffusion component of the PDE. Must be able to handle inputs with dimension (N,2) and output
-            arrays with dimension (N,2,2). I.e the N diffusion tensors at the given N 2D input locations. We note that
-            the diffusion tensor must have non-negative characteristic form.
+                arrays with dimension (N,2,2). I.e the N diffusion tensors at the given N 2D input locations. We note
+                that the diffusion tensor must have non-negative characteristic form.
             reaction: The reaction component of the PDE. Must be able to handle inputs with dimension (N,2) and output
-            arrays with dimension (N,). I.e. the N reaction coefficients at the given N 2D input locations.
+                arrays with dimension (N,). I.e. the N reaction coefficients at the given N 2D input locations.
             forcing: The forcing component of the PDE. Must be able to handle inputs with dimension (N,2) and output
-            arrays with dimension (N,). I.e. the N forcing values at the given N 2D input locations.
+                arrays with dimension (N,). I.e. the N forcing values at the given N 2D input locations.
             dirichlet_bcs: The Dirichlet boundary conditions associated with the PDE. Must be able to handle inputs
-            with dimension (N,2) and output arrays with dimension (N,). I.e. the N boundary values at the given N 2D
-            input locations. This can be mixed with relavent Neumann boundary conditions or stand-alone.
+                with dimension (N,2) and output arrays with dimension (N,). I.e. the N boundary values at the given N 2D
+                input locations. This can be mixed with relavent Neumann boundary conditions or stand-alone.
 
         Raises:
[…]
