+++ docs/requirements.txt
@@ -1,0 +1,2 @@
+sphinx>=7.4.7
+sphinx-rtd-theme~=3.0.2

[…]
