+++ docs/requirements.txt
@@ -1,2 +1,3 @@
 sphinx>=7.4.7
 sphinx-rtd-theme~=3.0.2
+myst-parser>=3.0.1
