+++ tests/test_DGFEM.py
@@ -1,0 +1,515 @@
+import pytest
+import numpy as np
+
+from reyna.polymesher.two_dimensional.domains import RectangleDomain, CircleDomain, RectangleCircleDomain, HornDomain
+from reyna.polymesher.two_dimensional.main import poly_mesher
+
+from reyna.geometry.two_dimensional.DGFEM import DGFEMGeometry
+
+from reyna.DGFEM.two_dimensional.main import DGFEM
+
+
+class TestDGFEM:
+
+    # Parameter tests to begin
+
+    def test_no_advection_or_diffusion(self):
+        """Test the failure when neither advection nor diffusion are used. """
+
+        dom = CircleDomain(np.array([[0, 1], [0, 1]]))
+        mesh = poly_mesher(dom, n_points=32)
+
+        geometry = DGFEMGeometry(mesh)
+        dg = DGFEM(geometry)
+
+        with pytest.raises(ValueError):
+            dg.add_data(dirichlet_bcs=lambda x: np.zeros(x.shape[0]))
+
+    def test_dirichlet_bcs_input(self):
+        """ Test is the bcs input is functional as expected when no boundary conditions are inputted. """
+
+        dom = CircleDomain(np.array([[0, 1], [0, 1]]))
+        mesh = poly_mesher(dom, n_points=32)
+
+        geometry = DGFEMGeometry(mesh)
+        dg = DGFEM(geometry)
+
+        advection = lambda x: np.ones(x.shape, dtype=float)
+
+        with pytest.raises(ValueError):
+            dg.add_data(advection=advection)
+
+    def test_diffusion_polynomial_degree_warning(self):
+        """ Test to see if the warning and correction is made when the polynomial degree is set too low. """
+
+        dom = RectangleCircleDomain()
+        mesh = poly_mesher(dom, n_points=128)
+
+        geometry = DGFEMGeometry(mesh)
+        dg = DGFEM(geometry, polynomial_degree=0)
+
+        def diffusion(x):
+            out = np.zeros((x.shape[0], 2, 2), dtype=np.float64)
+            for i in range(x.shape[0]):
+                out[i, 0, 0] = 1.0
+                out[i, 1, 1] = 1.0
+            return out
+
+        with pytest.warns(UserWarning, match='The polynomial degree given is too low'):
+            dg.add_data(diffusion=diffusion, dirichlet_bcs=lambda x: np.zeros(x.shape[0]))
+
+    def test_verbose_setting(self):
+        """ Test to see if the verbose counter fails as expected. """
+
+        dom = RectangleDomain(np.array([[0, 1], [0, 1]]))
+        mesh = poly_mesher(dom, n_points=16)
+
+        geometry = DGFEMGeometry(mesh)
+        dg = DGFEM(geometry)
+
+        def diffusion(x):
+            out = np.zeros((x.shape[0], 2, 2), dtype=np.float64)
+            for i in range(x.shape[0]):
+                out[i, 0, 0] = 1.0
+                out[i, 1, 1] = 1.0
+            return out
+
+        dg.add_data(diffusion=diffusion, dirichlet_bcs=lambda x: np.zeros(x.shape[0]))
+
+        with pytest.raises(AssertionError):
+            dg.dgfem(verbose=2)
+
+    def test_advection_constant_solution_polynomial_degree_one(self):
+        """ Test to see if constant solutions are captured exactly for advection equations with p=1 polynomials. """
+
+        dom = RectangleDomain(np.array([[0, 1], [0, 1]]))
+        mesh = poly_mesher(dom, n_points=16)
+
[…]
