+++ requirements.txt
@@ -3,3 +3,4 @@ scipy >= 1.4.0
 shapely >= 2.0.0
 matplotlib >= 3.4.0
 numba~=0.60.0
+pytest~=8.4.2

[…]
