### 2025-11-09 bb0f8ec
Completed the testing plus fixed a small bug in the DGFEM solver.
