### 2025-11-09 f53c418
added teh few final details and yml file for the automated github testing
