### 2025-11-09 dfae89c
Updated the tests.yml file to include numpy installation
