+++ .github/workflows/tests.yml
@@ -18,7 +18,9 @@ jobs:
 
       - name: Install dependencies
         run: |
-          pip install pytest
+          python -m pip install --upgrade pip
+          pip install -r requirements.txt
 
       - name: Run tests
-        run: pytest
+        run: |
+          pytest
