### 2025-12-19 b295d1b
Updated the documentation to reflect JOSS comments as well as cleaned up several parts of the testing and main scripts.
