+++ reyna/polymesher/two_dimensional/_auxilliaries/distance_functions.py
@@ -1,6 +1,5 @@
 import numpy as np
 import typing
-import matplotlib.pyplot as plt
 
 
 def d_sphere(p: np.ndarray, center: typing.Optional[np.ndarray] = None, radius: float = 1.0) -> np.ndarray:
@@ -12,14 +11,14 @@ def d_sphere(p: np.ndarray, center: typing.Optional[np.ndarray] = None, radius: 
 
     Args:
         p (np.ndarray): Array of points of the form `[p_0, p_1, ..., p_n]` where each `p_i` is `[x_i, y_i]` or
-        `[x_i, y_i, z_i]`.
+                        `[x_i, y_i, z_i]`.
         center (np.ndarray, optional): The center of the spherical domain. Defaults to the origin of the corresponding
-        dimension if `None`.
+                                       dimension if `None`.
         radius (float, optional): Radius of the spherical domain. Defaults to `1.0`.
 
     Returns:
         np.ndarray: Array of signed distances in the form `[d_0, d_1, ..., d_n]`,
-        where negative values indicate points inside the sphere and positive values outside.
+                    where negative values indicate points inside the sphere and positive values outside.
 
     Raises:
         Exception: p must contain points of dimention 2 or 3.
@@ -69,22 +68,24 @@ def d_rectangle(p: np.ndarray, x1: float, x2: float, y1: float, y2: float) -> np
 
 
 def d_line(p: np.ndarray, x1: float, y1: float, x2: float, y2: float) -> np.ndarray:
-    """Calculates signed distances from points to a line.
+    """
+    Calculates signed distances from points to a line.
 
-        Given an array of 2D coordinates `p`, this function computes the signed distance to a given line. This is a
-        constructor object and can be used to build more complicated domains. Defined with two points which the line
-        passes through. The interior side is defined as the positive normal.
+    Given an array of 2D coordinates `p`, this function computes the signed distance to a given line. This is a
+    constructor object and can be used to build more complicated domains. Defined with two points which the line
+    passes through. The interior side is defined as the positive normal.
 
-        Args:
-            p (np.ndarray): Array of 2D points.
-            x1 (float): First points x value.
-            y1 (float): First points y value.
-            x2 (float): Second points x value.
-            y2 (float): Second points y value.
+    Args:
+        p (np.ndarray): Array of 2D points.
+        x1 (float): First points x value.
+        y1 (float): First points y value.
+        x2 (float): Second points x value.
+        y2 (float): Second points y value.
 
-        Returns:
-            np.ndarray: Array of signed distances.
-        """
+    Returns:
+        np.ndarray: Array of signed distances.
+
+    """
     # tangent vector
     a = np.array([x2 - x1, y2 - y1])
     a = a/np.linalg.norm(a)
@@ -100,6 +101,21 @@ def d_line(p: np.ndarray, x1: float, y1: float, x2: float, y2: float) -> np.ndar
 
 
 def d_hexagon(p: np.ndarray, x: float = 0.0, y: float = 0.0, scale: float = 1.0) -> np.ndarray:
+    """
+    Calculates signed distances from points to a specified hexagon.
+
+    Given an array of 2D coordinates `p`, this function computes the signed distance to a given hexagon.
+
+    Args:
+        p (np.ndarray): Array of 2D points.
+        x (float): x value of the centre of the hexagon.
+        y (float): y value of the centre of the hexagon.
+        scale (float): the scale of the hexagon.
+
+    Returns:
+        np.ndarray: Array of signed distances.
+
+    """
     vertices = np.array([
[…]
