### 2025-12-19 ea7b853
Added the ```evaluate``` function and updated the documentation and README.md to reflect this
