### 2025-12-19 mattevs24/reyna#1 @mattevs24
@meliao An update has been provided to expose this useful method to users. This is now available in ```reyna.DGFEM.two_dimensional.tools``` alongside the ```plot_DG``` function. This has replaced the ```plotter``` module.
