### 2026-01-02 4144526
Updated the README.md file to reflect JOSS comments. Specifically, point to the upcoming features (Neumann boundary conditions as well as the 3D implementation).
