### 2026-01-02 #9292 @mattevs24
Hi @baxmittens,

Thank you very much for the reviewing work. We really appreciate the time spent on this by all reviewers and editors.

Thank you for your closing remarks: The second two have been addressed in the ```README.md``` file in the upcoming features and installation sections respectively. More examples will be created and uploaded over the coming year -- we're hoping to add more functionality, and more examples will be able to show this off!

Best wishes,

Matt
