### 2026-01-02 3072a58
Enhance README with project badges
Added badges for version, downloads, license, documentation, tests, and GitHub issues.
