### 2026-01-08 f32208b
Fix formatting and update DOI entries in references.bib
