+++ paper/references.bib
@@ -87,18 +87,8 @@
   volume={5},
   number={1},
   pages={21--61},
-  year={2017}
-}
-
-@article{Dedner:2017,
-  title={The DUNE-FEM-DG module.},
-  author={Dedner, Andreas and Girke, Stefan and Kl{\"o}fkorn, Robert and Malkmus, Tobias},
-  journal={Archive of Numerical Software},
-  volume={5},
-  number={1},
-  pages={21--61},
   year={2017},
-  doi={10.11588/ans.2017.1.28602}
+  doi={10.11588/ans.2017.1.28602},
 }
 
 @article{Badia:2018,
@@ -110,7 +100,7 @@
   pages={195--271},
   year={2018},
   publisher={Springer},
-  doi={10.1007/s11831-017-9244-1}
+  doi={10.1007/s11831-017-9244-1},
 }
 
 @article{Anderson:2021,
@@ -121,7 +111,7 @@
   pages={42--74},
   year={2021},
   publisher={Elsevier},
-  doi={https://doi.org/10.1016/j.camwa.2020.06.009},
+  doi={10.1016/j.camwa.2020.06.009},
 }
 
 @inproceedings{Prud:2012,
@@ -132,7 +122,7 @@
   pages={429--455},
   year={2012},
   organization={EDP Sciences},
-  doi={10.1051/proc/201238024}
+  doi={10.1051/proc/201238024},
 }
 
 @article{Bangerth:2007,
@@ -144,7 +134,7 @@
   pages={24--es},
   year={2007},
   publisher={ACM New York, NY, USA},
-  doi={10.1145/1268776.1268779}
+  doi={10.1145/1268776.1268779},
 }
 
 @article{alnaes:2015,
@@ -154,7 +144,7 @@
   volume={3},
   number={100},
   year={2015},
-  doi={10.11588/ans.2015.100.20553}
+  doi={10.11588/ans.2015.100.20553},
 }
 
 @article{Rathgeber:2016,
@@ -166,16 +156,16 @@
   pages={1--27},
   year={2016},
   publisher={ACM New York, NY, USA},
-  doi={10.1145/2998441}
+  doi={10.1145/2998441},
 }
 
 @article{Ching:2022,
-title = {Quail: A lightweight open-source discontinuous Galerkin code in Python for teaching and prototyping},
-journal = {SoftwareX},
-volume = {17},
-pages = {100982},
-year = {2022},
-issn = {2352-7110},
-doi = {https://doi.org/10.1016/j.softx.2022.100982},
-author = {Eric J. Ching and Brett Bornhoft and Ali Lasemi and Matthias Ihme},
+  title = {Quail: A lightweight open-source discontinuous Galerkin code in Python for teaching and prototyping},
+  journal = {SoftwareX},
+  volume = {17},
+  pages = {100982},
+  year = {2022},
+  issn = {2352-7110},
+  doi = {10.1016/j.softx.2022.100982},
+  author = {Eric J. Ching and Brett Bornhoft and Ali Lasemi and Matthias Ihme},
 }
