### 2026-01-12 13e11c4
began some of the poly_mesher work -- moved over a lot of the previous work, will need to refine to match the performance and style of the ```polymesher.two_dimensional.main``` code.
