+++ reyna/polymesher/three_dimensional/main.py
@@ -1,0 +1,239 @@
+import numpy as np
+from scipy.spatial import Voronoi, Delaunay
+
+from reyna.polymesher.three_dimensional._auxilliaries.abstraction import Domain3D, PolyMesh3D
+
+
+def poly_mesher_3d(domain: Domain3D, max_iterations: int = 100, **kwargs) -> PolyMesh3D:
+    """
+
+    :param domain:
+    :param max_iterations:
+    :param kwargs:
+    :return:
+    """
+    if "n_points" in kwargs:
+        n_points = kwargs.get("n_points")
+        points = poly_mesher_init_point_set(domain, n_points=n_points)
+
+    elif "n_x" in kwargs and "n_y" in kwargs and "n_z" in kwargs:
+        n_x = kwargs.get("n_x")
+        n_y = kwargs.get("n_y")
+        n_z = kwargs.get("n_z")
+        points = poly_mesher_init_point_set(domain, n_x=n_x, n_y=n_y, n_z=n_z)
+
+    else:
+        raise AttributeError("key word error: must be just `n_points` or both `n_x` and `n_y`.")
+
+    fixed_points = domain.pFix()  # from here can call domain.fixed_points -- this initialises the property simult
+    if fixed_points is not None:
+        points = np.concatenate((fixed_points, points), axis=0)
+        n_fixed = fixed_points.shape[0]
+    else:
+        n_fixed = 0
+
+    iteration, error, tolerance = 0, 1.0, 1e-4
+    area = domain.volume()
+    n_points = points.shape[0]
+
+    voronoi, nodes, elements = None, None, None
+
+    while iteration <= max_iterations and error > tolerance:
+        reflected_points = poly_mesher_reflect(points, domain, area)
+
+        voronoi = Voronoi(np.concatenate((points, reflected_points), axis=0), qhull_options='Qbb Qz')
+
+        cond_len = len(voronoi.regions[0]) == 0
+
+        nodes = voronoi.vertices
+        if cond_len:
+            sorting = [np.where(voronoi.point_region == x)[0][0] for x in range(1, len(voronoi.regions))]
+            elements = [x for _, x in sorted(zip(sorting, voronoi.regions[1:]))]
+        else:
+            empty_ind = voronoi.regions.index([])
+            sorting_1 = [np.where(voronoi.point_region == x)[0][0] for x in range(0, empty_ind)]
+            sorting_2 = [np.where(voronoi.point_region == x)[0][0] for x in range(empty_ind+1, len(voronoi.regions))]
+            sorting = sorting_1 + sorting_2
+            regions = voronoi.regions
+            del regions[empty_ind]
+            elements = [x for _, x in sorted(zip(sorting, regions))]
+
+        points, area, error = poly_mesher_vorocentroid(points, nodes, elements)
+
+        if fixed_points is not None:
+            points[:n_fixed, :] = fixed_points
+
+        iteration += 1
+        # show_mesh(voronoi.vertices, elements[: n_points], bounding_box=bounding_box, alpha=0.4, edgecolor="#556575")
+
+        if iteration % 10 == 0:
+            print(f"Iteration: {iteration}. Error: {error}")
+
+    poly_mesh = PolyMesh3D(voronoi.vertices, elements[: n_points], points, domain)
+
+    return poly_mesh
+
+
+def poly_mesher_init_point_set(domain: Domain3D, **kwargs) -> np.ndarray:
+    """
+    This function initialises the point set to be able to use the lloyds algorithm. Can be done uniformly via ```n_x```
+    and ```n_y``` or the number of points that reside in the domain via ```n_points```.
[…]
