### 2026-01-13 033a5ea
First working update for polymesher in 3D! Functionality is minimal and there are bugs to fix but these will be availible in a future update.
