### 2026-01-14 15fec8c
Added the ```three_dimensional``` geometry object and began to populate
