+++ reyna/geometry/three_dimensional/DGFEM.py
@@ -1,0 +1,156 @@
+import time
+import typing
+
+import numpy as np
+from scipy.spatial import Delaunay
+from scipy.sparse import csr_matrix, find
+from shapely import Polygon, Point
+
+from reyna.polymesher.three_dimensional._auxilliaries.abstraction import PolyMesh3D
+
+from dataclasses import dataclass
+import pickle
+
+
+@dataclass
+class DGFEMGeometry:
+    """
+    This is a geometry function which provides all the additional mesh information that a DGFEM method requires to
+    run; normal vectors, subtriagulations/simplicial decompositions etc.
+    """
+
+    # TODO: this will be a lot more complicated than the two dimensional version. For example here, we need information
+    #  on all the facets as well as the elements themselves. This will make this function more costly but hopefully it
+    #  won't be prhibitive as it wasn't up to a lot of elements in the two dimensional case.
+
+    def __init__(self, poly_mesh: PolyMesh3D, **kwargs):
+        """
+        Initializes DGFEMGeometry with the given PolyMesh object.
+
+        Args:
+            poly_mesh (PolyMesh): The polygonal mesh of which to generate the information.
+            **kwargs: Additional keyword arguments. 'time' is the only current option and is used to time the geometry
+            generation.
+        """
+        self.mesh = poly_mesh
+
+        self.n_nodes = poly_mesh.vertices.shape[0]
+        self.n_elements = len(poly_mesh.elements)
+
+        self.nodes = poly_mesh.vertices
+
+        self.elem_bounding_boxes = None  # yes -- done
+
+        self.boundary_facets = None  # yes
+        self.boundary_facets_to_element = None  # yes
+        self.interior_facets = None  # yes
+        self.interior_facets_to_element = None  # yes
+        self.boundary_normals = None  # yes
+        self.interior_normals = None  # yes
+
+        self.simplicial_decomposition = None  # yes -- done
+        self.facet_subtriangulations = None  # yes -- done
+        self.n_simplicies = None  # yes -- done
+        self.simplex_to_element = None  # yes -- done
+        self.triangle_to_facet = None   # yes? -- done
+
+        self.h: typing.Optional[float] = None  # yes
+        self.h_s: typing.Optional[np.ndarray] = None  # yes
+        self.volumes: typing.Optional[np.ndarray] = None  # yes
+
+        # TODO: I may need omre attributes than this but not 100% sure currently.
+
+        time_generation = False
+        if 'time' in kwargs:
+            time_generation = kwargs.pop('time')
+
+        _time = time.time()
+
+        self._generate()
+
+        if time_generation:
+            print(f"Time taken to generate geometry: {time.time() - _time}s")
+
+    def _generate(self, max_n: int = 100):
+
+        self.nodes *= 1e8
+        self.nodes /= 1e8
+
+        sub_simplicial_decomp_0 = []
+        sub_simplicial_decomp_1 = []
+        sub_simplicial_decomp_2 = []
+        sub_simplicial_decomp_3 = []
+
+        simplex_to_element = []
+        elem_bounding_boxes = []
+
+        # Operations on the elements
+
+        for i, element in enumerate(self.mesh.elements):
+
+            element = np.array(element)
+
+            # Bounding box calculation here
+            elem_bounding_boxes.append([np.min(self.nodes[element, 0]), np.max(self.nodes[element, 0]),
+                                        np.min(self.nodes[element, 1]), np.max(self.nodes[element, 1]),
+                                        np.min(self.nodes[element, 2]), np.max(self.nodes[element, 2])])
+
+            # Simplicial Decomposition here
+
+            local_simplicial_decomp = Delaunay(self.nodes[element, :])
[…]
