### 2026-01-14 d18408d
Added lots of work to the 3D solver -- need to match up DGFEMGeometry and DGFEM nicely -- heavy refactoring is required, particularly with the focus on combining both 2- and 3d solvers into one.
