+++ reyna/DGFEM/three_dimensional/main.py
@@ -1,0 +1,645 @@
+import typing
+import time
+import warnings
+
+import numpy as np
+from numba import njit, f8
+from scipy.sparse import csr_matrix
+from scipy.sparse.linalg import spsolve
+
+from numpy.polynomial.legendre import leggauss
+from scipy.special import roots_jacobi
+
+from reyna.geometry.two_dimensional.DGFEM import DGFEMGeometry
+
+from reyna.DGFEM.two_dimensional._auxilliaries.boundary_information import BoundaryInformation
+from reyna.DGFEM.three_dimensional._auxilliaries.polyhedral_basis_utils import basis_indices
+
+from reyna.DGFEM.three_dimensional._auxilliaries.assembly.full_assembly import localstiff, int_localstiff
+
+from reyna.DGFEM.two_dimensional._auxilliaries.assembly.boundary_assembly import (local_advection_inflow,
+                                                                                  local_diffusion_dirichlet)
+
+from reyna.DGFEM.two_dimensional._auxilliaries.error_assembly.boundary_assembly import error_d_face, error_a_face
+from reyna.DGFEM.two_dimensional._auxilliaries.error_assembly.full_assembly import error_element, error_interface
+
+from reyna.DGFEM.two_dimensional.tools import plot_DG
+
+
+class DGFEM:
+    """
+    This class contains all the methods required for the DGFEM solver.
+
+    Attributes:
+        geometry (DGFEMGeometry): The geometry associated with the computation domain for the problem.
+        h (float): The maximal cell diameter of the mesh.
+        polydegree (int): The highest total degree polynomial space required.
+
+        advection (typing.Optional[typing.Callable[[np.ndarray], np.ndarray]]): The advection coeffient.
+        diffusion (typing.Optional[typing.Callable[[np.ndarray], np.ndarray]]): The diffusion coeffient.
+        reaction (typing.Optional[typing.Callable[[np.ndarray], np.ndarray]]): The reaction coeffient.
+        forcing (typing.Optional[typing.Callable[[np.ndarray], np.ndarray]]): The forcing term.
+        dirichlet_bcs (typing.Optional[typing.Callable[[np.ndarray], np.ndarray]]): The dirichlet boundary conditions.
+
+        boundary_information (typing.Optional[BoundaryInformation]): The information associated with the boundary of the
+            computational domain and PDE coefficients.
+        sigma_D (float): The global penalisation coefficient.
+
+        orders (typing.Optional[np.ndarray]): The orders of the tensor-Legendre polynomials.
+
+        dim_elem (typing.Optional[int]): The local finite element space dimension.
+        dim_system (typing.Optional[int]): The global system dimension.
+
+        solution (typing.Optional[np.ndarray]): The solution vector for the coefficients of the Legendre polynomials.
+
+        B (typing.Optional[csr_matrix]): The global stiffness matrix (populated when .dgfem() method is called).
+        L (typing.Optional[np.ndarray]): The global forcng vector (populated when .dgfem() method is called).
+
+        element_reference_quadrature (typing.Optional[typing.Tuple[np.ndarray, np.ndarray]]): The 2d reference
+            quadrature rule.
+        edge_reference_quadrature (typing.Optional[typing.Tuple[np.ndarray, np.ndarray]]): The 1d reference quadrature
+            rule.
+
+    Methods:
+        add_data(...): Add the function coefficients for the non-negative characteristic second order PDE.
+        dgfem(...): Run the solver with the information and geometry inputted.
+        errors(...): Given the numerically solved PDE, find the errors associated with the numerical solution.
+        plot_DG(...): Plot the DGFEM solution.
+
+    See Also:
+        - See the reference and example notebooks for examples on how this class is used fully.
+    """
+
[…]
