### 2026-01-14 d3956d3
Updates to DGFEM solver and the surrounding files -- generality where possible
