+++ reyna/polymesher/three_dimensional/main.py
@@ -71,9 +71,15 @@ def poly_mesher_3d(domain: Domain3D, max_iterations: int = 100, **kwargs) -> Pol
         if iteration % 10 == 0:
             print(f"Iteration: {iteration}. Error: {error}")
 
-    filtered_facets = _filter_facets(voronoi.ridge_points, voronoi.ridge_vertices, n_points)
+    facet_types = np.sum(voronoi.ridge_points < n_points, axis=1)
+    mask = facet_types > 0
+    filtered_facets = list(filter(lambda x: mask[voronoi.ridge_vertices.index(x)], voronoi.ridge_vertices))
 
-    poly_mesh = PolyMesh3D(voronoi.vertices, filtered_facets, elements[: n_points], points, domain)
+    poly_mesh = PolyMesh3D(
+        voronoi.vertices, filtered_facets, elements[: n_points],
+        facet_types[mask] - 1, voronoi.ridge_points[mask],
+        points, domain
+    )
     # TODO: need to be careful -- can use ridge_points and ridge_vertices but have to restrict in some way to include
     #  the boundary facets too -- if a row in ridge points contains an index less than n_points, retain and filter
     #  ridge vertices with this (the indecies should be the global ones?) this could optimise the boundary edge and
@@ -244,20 +250,3 @@ def _3d_simplex_volume(vertices: np.ndarray) -> float:
     ad = vertices[3, :] - vertices[0, :]
     volume = np.abs(1.0 / 6.0 * np.linalg.det(np.vstack((ab, ac, ad))))
     return volume
-
-
-def _filter_facets(ridge_points: np.ndarray, ridge_vertices: typing.List[list], n_elements: int) -> typing.List[list]:
-    """
-    This function filters out the facets that correpond to the bounded Voronoi tessellation.
-    Args:
-        ridge_points: Indices of the points between which each Voronoi ridge lies.
-        ridge_vertices: Indices of the Voronoi vertices forming each Voronoi ridge.
-        n_elements: Number of points or elements in the bounded mesh.
-
-    Returns:
-        (typing.List[list]): List of lists of integer indecies for 'vertices'
-    """
-    mask = np.sum(ridge_points < n_elements, axis=1) > 0
-    filtered_facets = list(filter(lambda x: mask[ridge_vertices.index(x)], ridge_vertices))
-
-    return filtered_facets
