### 2026-01-22 #9292 @drvinceknight
Ok I don't have any further requests for the paper, it all looks good. Can you make a tagged release and a new archive, and report the version number and archive DOI here please.
