### 2026-01-22 #9292 @mattevs24
Hi @drvinceknight,

Thank you very much for all the work gone into this.

The release is tagged under v1.4.0 and the zenodo archive has the DOI: https://doi.org/10.5281/zenodo.18339335.

Matt
