+++ paper/references.bib
@@ -1,5 +1,5 @@
 @article{Alberty:1999,
-  title={Remarks around 50 lines of Matlab: short finite element implementation},
+  title={Remarks around 50 lines of {M}atlab: short finite element implementation},
   author={Alberty, Jochen and Carstensen, Carsten and Funken, Stefan A},
   journal={Numerical algorithms},
   volume={20},
@@ -23,7 +23,7 @@
 }
 
 @article{Calloo:2025,
-  title={Cycle-free polytopal mesh sweeping for Boltzmann transport},
+  title={Cycle-free polytopal mesh sweeping for {B}oltzmann transport},
   author={Calloo, Ansar and Evans, Matthew and Lockyer, Henry and Madiot, Fran{\c{c}}ois and Pryer, Tristan and Zanetti, Luca},
   journal={Numerical Algorithms},
   pages={1--24},
@@ -33,7 +33,7 @@
 }
 
 @article{Cangiani:2014,
-  title={hp-version discontinuous Galerkin methods on polygonal and polyhedral meshes},
+  title={{hp}-version discontinuous {G}alerkin methods on polygonal and polyhedral meshes},
   author={Cangiani, Andrea and Georgoulis, Emmanuil H and Houston, Paul},
   journal={Mathematical Models and Methods in Applied Sciences},
   volume={24},
@@ -45,7 +45,7 @@
 }
 
 @article{Cangiani:2016,
-  title={hp-version discontinuous Galerkin methods for advection-diffusion-reaction problems on polytopic meshes},
+  title={{hp}-version discontinuous {G}alerkin methods for advection-diffusion-reaction problems on polytopic meshes},
   author={Cangiani, Andrea and Dong, Zhaonan and Georgoulis, Emmanuil H and Houston, Paul},
   journal={ESAIM: Mathematical Modelling and Numerical Analysis},
   volume={50},
@@ -57,7 +57,7 @@
 }
 
 @article{Talischi:2012,
-  title={PolyMesher: a general-purpose mesh generator for polygonal elements written in Matlab},
+  title={PolyMesher: a general-purpose mesh generator for polygonal elements written in {M}atlab},
   author={Talischi, Cameron and Paulino, Glaucio H and Pereira, Anderson and Menezes, Ivan FM},
   journal={Structural and Multidisciplinary Optimization},
   volume={45},
@@ -115,7 +115,7 @@
 }
 
 @inproceedings{Prud:2012,
-  title={Feel++: A computational framework for galerkin methods and advanced numerical methods},
+  title={Feel++: A computational framework for {G}alerkin methods and advanced numerical methods},
   author={Prud’Homme, Christophe and Chabannes, Vincent and Doyeux, Vincent and Ismail, Mourad and Samake, Abdoulaye and Pena, Gon{\c{c}}alo},
   booktitle={ESAIM: Proceedings},
   volume={38},
@@ -126,7 +126,7 @@
 }
 
 @article{Bangerth:2007,
-  title={deal. II—a general-purpose object-oriented finite element library},
+  title={deal.II—a general-purpose object-oriented finite element library},
   author={Bangerth, Wolfgang and Hartmann, Ralf and Kanschat, Guido},
   journal={ACM Transactions on Mathematical Software (TOMS)},
   volume={33},
@@ -160,7 +160,7 @@
 }
 
 @article{Ching:2022,
-  title = {Quail: A lightweight open-source discontinuous Galerkin code in Python for teaching and prototyping},
+  title = {Quail: A lightweight open-source discontinuous {G}alerkin code in {P}ython for teaching and prototyping},
   journal = {SoftwareX},
   volume = {17},
   pages = {100982},
