### 2026-02-03 #9292 @mattevs24
@danielskatz This looks suitable to me, PR accepted and merged!
