### 2026-02-03 #9292 @mattevs24
@danielskatz, my only suggestion is with the reviewers -- as @meliao suggested;

> [@drvinceknight](https://github.com/drvinceknight) , I think [@lamBOOO](https://github.com/lamBOOO) should be removed from the reviewer list as they appear in the article proof under "reviewers". Sorry for noticing this so late in the process.

The corrections in the PR are fine, and I think they tidy up the paper nicely. The corrections to the bibliography entries are also good! Thank you very much for your time!
