### 2026-02-03 #9292 @danielskatz
Congratulations to @mattevs24 (Matthew Evans) and co-author on your JOSS publication!!

JOSS completely depends on volunteers - we couldn't do this without you

(At least for me, the PDF isn't currently loading, so I'll keep this open until is does. This is typically a network caching issue)
